+++ backend/LawsLaboratory/LawsLaboratory.Core/SpatialModel/VariableReference.cs
@@ -1,6 +1,6 @@
 ﻿namespace LawsLaboratory.Core.SpatialModel.Position;
 
 public readonly record struct VariableReference(
-    int ParameterId,
+    ushort ParameterId,
     PlanePosition RelativePosition
     );
